pyproject.toml
[project]
name = "personal-assistant"
version = "0.1.0"
description = "Telegram personal assistant that creates Google Calendar events via OpenAI"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot>=21.0",
    "openai>=1.30.0",
    "google-api-python-client>=2.130.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pytz>=2024.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
